splitwise-worker/pyproject.toml
[project]
name = "splitwise-worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "boto3>=1.42.61",
    "mypy-boto3-s3>=1.42.37",
    "mypy-boto3-sqs>=1.42.3",
    "pandas>=3.0.1",
    "python-decouple>=3.8",
    "python-json-logger>=4.0.0",
]
